Use 64bit diff throughout
This appears to be mainly cosmetic. Difficulty values from the daemon
are larger than 32bits, and the value being displayed was being truncated.
Doesn't appear to affect mining, which uses the 64bit target.

+++ src/api/NetworkState.h
@@ -50,7 +50,7 @@ public:
 
     char pool[256];
     std::array<uint64_t, 10> topDiff { { } };
-    uint32_t diff;
+    uint64_t diff;
     uint64_t accepted;
     uint64_t failures;
     uint64_t rejected;
